remove sporco from bench_methods and merge with dicodile version
