+++ benchmarks/bench_methods_run.py
@@ -3,9 +3,8 @@ Benchmark different solver of the same CSC univariate or multivariate problem.
 
 This script needs the following packages:
     pip install pandas pyfftw
-    pip install alphacsc/other/sporco
 
-- Use bench_methods_run.py to run the benchmark.
+- Use bench_methods_proposed_run.py to run the benchmark.
   The results are saved in alphacsc/figures.
 - Use bench_methods_plot.py to plot the results.
   The figures are saved in alphacsc/figures.
@@ -20,9 +19,6 @@ import numpy as np
 import pandas as pd
 import scipy.sparse as sp
 from joblib import Parallel, delayed
-
-import alphacsc.other.heide_csc as CSC
-from sporco.admm.cbpdndl import ConvBPDNDictLearn
 
 from alphacsc.update_d import update_d_block
 from alphacsc.learn_d_z import learn_d_z
@@ -56,75 +52,6 @@ reg_list = [10.]
 ######################################
 # Functions compared in the benchmark
 ######################################
-
-
-def run_admm(X, ds_init, reg, n_iter, random_state, label, max_it_d=10,
-             max_it_z=10):
-    # admm with the following differences
-    # - positivity constraints
-    # - different init
-    # - d step and z step are swapped
-    tol = np.float64(1e-3)
-    size_kernel = ds_init.shape
-    assert size_kernel[1] % 2 == 1
-    [d, z, Dz, list_obj_val, times_admm] = CSC.learn_conv_sparse_coder(
-        X, size_kernel, max_it=n_iter, tol=tol, random_state=random_state,
-        lambda_prior=reg, ds_init=ds_init, verbose=verbose, max_it_d=max_it_d,
-        max_it_z=max_it_z)
-
-    # z.shape = (n_trials, n_atoms, n_times + 2 * n_times_atom)
-    z = z[:, :, 2 * n_times_atom:-2 * n_times_atom]
-    z = z.swapaxes(0, 1)
-    # z.shape = (n_atoms, n_trials, n_times - 2 * n_times_atom)
-
-    return list_obj_val, np.cumsum(times_admm)[::2], d, z
-
-
-def run_cbpdn(X, ds_init, reg, n_iter, random_state, label):
-    # Use only one thread in fft for fair comparison
-    import sporco.linalg
-    sporco.linalg.pyfftw_threads = 1
-
-    if X.ndim == 2:  # univariate CSC
-        ds_init = np.swapaxes(ds_init, 0, 1)[:, None, :]
-        X = np.swapaxes(X, 0, 1)[:, None, :]
-        single_channel = True
-    else:  # multivariate CSC
-        ds_init = np.swapaxes(ds_init, 0, 2)
-        X = np.swapaxes(X, 0, 2)
-        single_channel = False
-
-    options = {
-        'Verbose': verbose > 0,
-        'StatusHeader': False,
-        'MaxMainIter': n_iter,
-        'CBPDN': dict(NonNegCoef=True),
-        'CCMOD': dict(ZeroMean=False),
-        'DictSize': ds_init.shape,
-    }
-
-    # wolberg / convolutional basis pursuit
-    opt = ConvBPDNDictLearn.Options(options)
-    cbpdn = ConvBPDNDictLearn(ds_init, X, reg, opt, dimN=1)
-    results = cbpdn.solve()
-    times = np.cumsum(cbpdn.getitstat().Time)
-
-    d_hat, pobj = results
-    if single_channel:  # univariate CSC
-        d_hat = d_hat.squeeze().T
-        n_atoms, n_times_atom = d_hat.shape
-    else:
-        d_hat = d_hat.squeeze().swapaxes(0, 2)
-        n_atoms, n_channels, n_times_atom = d_hat.shape
-
-    z_hat = cbpdn.getcoef().squeeze().swapaxes(0, 2)
-    times = np.concatenate([[0], times])
-
-    # z_hat.shape = (n_atoms, n_trials, n_times)
-    z_hat = z_hat[:, :, :-n_times_atom + 1]
-    # z_hat.shape = (n_atoms, n_trials, n_times_valid)
-
-    return pobj, times, d_hat, z_hat
 
 
 def run_fista(X, ds_init, reg, n_iter, random_state, label):
@@ -194,6 +121,25 @@ def run_multichannel_gcd_fullrank(X, ds_init, reg, n_iter, random_state,
     return pobj[::2], np.cumsum(times)[::2], d_hat, z_hat
 
 
+def run_multichannel_dicodile_fullrank(X, ds_init, reg, n_iter, random_state,
+                                       label):
+    assert X.ndim == 3
+    n_atoms, n_channels, n_times_atom = ds_init.shape
+
+    solver_z_kwargs = dict(max_iter=1000, tol=1e-3)
+    pobj, times, d_hat, z_hat, reg = learn_d_z_multi(
+        X, n_atoms, n_times_atom, solver_d='auto', solver_z="dicodile",
+        uv_constraint='auto', eps=-np.inf, solver_z_kwargs=solver_z_kwargs,
+        reg=reg, solver_d_kwargs=dict(max_iter=100), n_iter=n_iter,
+        random_state=random_state, raise_on_increase=False, D_init=ds_init,
+        n_jobs=10, verbose=verbose, rank1=False)
+
+    # remove the ds init duration
+    times[0] = 0
+
+    return pobj[::2], np.cumsum(times)[::2], d_hat, z_hat
+
+
 def colorify(message, color=BLUE):
     """Change color of the standard output"""
     return ("\033[1;%dm" % color) + message + "\033[0m"
@@ -205,17 +151,17 @@ def colorify(message, color=BLUE):
 
 n_iter = 100
 methods_univariate = [
-    [run_cbpdn, 'Garcia-Cardona et al (2017)', n_iter * 2],
     [run_fista, 'Jas et al (2017) FISTA', n_iter],
     [run_l_bfgs, 'Jas et al (2017) LBFGS', n_iter],
-    [run_multichannel_gcd, 'Proposed (univariate)', n_iter],
+    [run_multichannel_gcd, 'gcd', n_iter],
 ]
 
 n_iter_multi = 20
 methods_multivariate = [
-    [run_cbpdn, 'Wohlberg (2016)', n_iter_multi * 2],
-    [run_multichannel_gcd_fullrank, 'Proposed (multivariate)', n_iter_multi],
-    [run_multichannel_gcd, 'Proposed (multichannel)', n_iter_multi],
+    [run_multichannel_gcd_fullrank, 'gcd fullrank', n_iter_multi],
+    [run_multichannel_dicodile_fullrank,
+        'dicodile fullrank', n_iter_multi],
+    [run_multichannel_gcd, 'gcd', n_iter_multi],
 ]
 
 
@@ -292,7 +238,7 @@ if __name__ == '__main__':
         all_results = []
 
         X, info = load_data(
-            dataset='somato', epoch=False, n_jobs=n_jobs, n_splits=2
+            dataset='somato', epoch=False, n_jobs=n_jobs, n_splits=1
         )
 
         if n_channels == 1:
@@ -300,7 +246,6 @@ if __name__ == '__main__':
         elif n_channels is not None:
             X = X[:, :n_channels, :]
 
-        assert X.shape[0] > 1  # we need at least two trials for sporco
         X_shape = X.shape
 
         if n_channels == 1:
